pyproject.toml
[project]
name = "ragsds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.16",
    "fastapi>=0.116.1",
    "huggingface-hub>=0.34.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.9",
    "langchain-huggingface>=0.3.1",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.3",
    "pypdf>=5.9.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
    "torch>=2.8.0",
    "transformers>=4.55.0",
    "uvicorn>=0.35.0",
]
